Fix type mismatch in calls to gsl_histogram(2d)?_find
The last argument(s) are size_t, which is an unsigned integer. The next
Rakudo version will finally have proper support for unsigned ints, which
means that it also knows the difference between int and uint and will
complain if given the wrong one. Furthermore on 64 bit systems size_t is
usually 64 bit large. Since the exact size is platform dependent, using
size_t directly as argument type is just a better idea anyway.

This is required for Rakudo 2022.02 and up. The fix is backwards
compatible.